pyproject.toml
[project]
name = "realtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "numpy>=2.3.2",
    "openai>=1.98.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
    "setuptools>=80.9.0",
    "webrtcvad>=2.0.10",
    "wheel>=0.45.1",
]
